Reload page when user logged out due to inactivity. Remove Home from Nav bar when on Home page

+++ src/App.jsx
@@ -40,14 +40,17 @@ function App() {
   const [data, setData] = useState('');
 
   const handleOnUserIdle = () => {
-    localStorage.clear();
-    setUsername(null);
-    setUserid(null);
-    setToken(null);
-    setProfilePicture(null);
+    if (localStorage.getItem('token')) {
+      localStorage.clear();
+      setUsername(null);
+      setUserid(null);
+      setToken(null);
+      setProfilePicture(null);
+      localStorage.getItem('token') === null && window.location.reload();
+    }
   }
 
-  const IDLE_TIME = 30 * 60 * 1000; // 30 mins in ms
+  const IDLE_TIME = 1 * 60 * 1000; // 30 mins in ms
   const GENERAL_DEBOUNCE_TIME = 500; // in ms
   useIdleTimer({
     timeout: IDLE_TIME,
